requirements.txt
aiofiles==24.1.0
aiohttp==3.11.18
aiosqlite==0.21.0
APScheduler==3.11.0
discord.py==2.5.2
Pillow==11.2.1
python-dotenv==1.1.0
Requests==2.32.3
